# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/despesa": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Lista despesas do mês atual",
                "operationId": "c07e6053bb5d89892d3633bc193d6813",
                "responses": {
                    "200": {
                        "description": "Lista de despesas com saldo disponível",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "finances": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Finance"
                                            }
                                        },
                                        "available_money": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AvailableMoney"
                                            }
                                        },
                                        "diff": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 250.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Finance"
                ],
                "summary": "Cria nova despesa",
                "operationId": "84f773edb67f0bd59219094a8884d8dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "value",
                                    "date",
                                    "available_money_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Compra supermercado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Compra semanal"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.75
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-28"
                                    },
                                    "available_money_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categories_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "payments_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "payable": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Despesa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Finance"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/despesa/{id}": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Mostra detalhes de uma despesa",
                "operationId": "e90278dab61b239945b93a481a7c21d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da despesa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Despesa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Finance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Despesa não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Finance"
                ],
                "summary": "Atualiza uma despesa",
                "operationId": "b43f988c072cd43c0509754fb0423bfa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da despesa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "value",
                                    "date",
                                    "available_money_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Compra supermercado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Compra semanal"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-10-01"
                                    },
                                    "available_money_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "categories_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "payments_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "payable": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Despesa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Finance"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Despesa não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Finance"
                ],
                "summary": "Deleta uma despesa",
                "operationId": "a9946cb97ebf1df9bac24eb0cbf748d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da despesa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Despesa deletada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Despesa deletada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Despesa não encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvailableMoney": {
                "title": "AvailableMoney",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "to_spend",
                    "date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Saldo inicial"
                    },
                    "to_spend": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-28"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cartão de crédito"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "required": [
                    "id",
                    "user_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Alimentação"
                    },
                    "color": {
                        "type": "string",
                        "example": "#FF0000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "Finance": {
                "title": "Despesa",
                "required": [
                    "id",
                    "name",
                    "value",
                    "date",
                    "available_money_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Compra supermercado"
                    },
                    "description": {
                        "type": "string",
                        "example": "Compra semanal de alimentos"
                    },
                    "value": {
                        "type": "number",
                        "format": "float",
                        "example": 150.75
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-28"
                    },
                    "payable": {
                        "type": "boolean",
                        "example": false
                    },
                    "available_money_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "categories_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "payments_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "payment": {
                        "$ref": "#/components/schemas/Payment"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Use o token de autenticação no formato Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}